refactor(mcp): centralize credential checks in credentials.py
